Refactor SeekerSchema
- Wanted to be clever and make the code more readable, but I think this form is more simple and ultimately better. Also I realized a lot of the default error messages cover the extra message I was trying to add.
- Also removed the JSON nesting that marshmallow was doing including a few utility functions that did that in favor of just returning a flat data structure to make the API more versatile to end-users who could choose how they want to group the data.
